# docs/swagger.yaml
basePath: /api/v1
definitions:
  circle.Circle:
    properties:
      created_at:
        description: Created at
        type: string
      created_by:
        description: Created by
        type: integer
      disabled:
        description: Disabled
        type: boolean
      expiredAt:
        description: read one column
        type: string
      id:
        description: Unique identifier
        type: integer
      invite_code:
        description: Invite code
        type: string
      name:
        description: Full name
        type: string
      subscriptionStatus:
        description: read one column
        type: string
      updated_at:
        description: Updated at
        type: string
      webhook_url:
        description: Webhook URL
        type: string
    type: object
...
  circle.UserCircleDetail:
    properties:
      circleId:
        description: Circle ID
        type: integer
      createdAt:
        description: Created at
        type: string
      displayName:
        type: string
      id:
        description: Unique identifier
        type: integer
      image:
        description: Image
        type: string
      isActive:
        type: boolean
      points:
        description: Points
        type: integer
      pointsRedeemed:
        description: Points Redeemed
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/circle.UserRole'
        description: Role
      updatedAt:
        description: Updated at
        type: string
      userId:
        description: User ID
        type: integer
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Donetick swagger documentation.
  license:
    name: GNU Affero General Public License v3.0
    url: https://github.com/donetick/donetick/blob/main/LICENSE.md
  title: Donetick Swagger API
  version: "1.0"
paths:
...
  /circles:
    get:
      consumes:
      - application/json
      description: Retrieves all circles the current user belongs to
      produces:
      - application/json
      responses:
        "200":
          description: 'res: array of circles'
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/circle.Circle'
              type: array
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error getting user circles'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Get user circles
      tags:
      - circles
  /circles/{id}/members/delete:
    delete:
      consumes:
      - application/json
      description: Removes a member from a circle (admin only)
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Member ID to delete
        in: query
        name: member_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'res: User deleted from circle successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: You are not an admin of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error deleting circle member'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Delete a circle member
      tags:
      - circles
  /circles/{id}/members/points/redeem:
    post:
      consumes:
      - application/json
      description: Redeems points for a circle member (admin only)
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Points redemption request
        in: body
        name: points
        required: true
        schema:
          properties:
            points:
              type: integer
            userId:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'res: Points redeemed successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request / User does not have enough points
            / User is not a member of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: You are not an admin of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error redeeming points'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Redeem member points
      tags:
      - circles
  /circles/join:
    post:
      consumes:
      - application/json
      description: Requests to join a circle using an invite code
      parameters:
      - description: Invite code
        in: query
        name: invite_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'res: User Requested to join circle successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 'error: You are already a member of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error adding user to circle'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Request to join a circle
      tags:
      - circles
  /circles/leave:
    delete:
      consumes:
      - application/json
      description: Removes the current user from a circle and restores their original
        circle
      parameters:
      - description: Circle ID
        in: query
        name: circle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'res: User left circle successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error leaving circle'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Leave a circle
      tags:
      - circles
  /circles/members:
    get:
      consumes:
      - application/json
      description: Retrieves all members of the current user's circle
      produces:
      - application/json
      responses:
        "200":
          description: 'res: array of circle members'
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/circle.UserCircleDetail'
              type: array
            type: object
        "500":
          description: 'error: Error getting circle members'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Get circle members
      tags:
      - circles
  /circles/members/requests:
    get:
      consumes:
      - application/json
      description: Retrieves pending circle join requests (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: 'res: array of pending members'
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/circle.UserCircleDetail'
              type: array
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: You are not an admin of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error getting pending circle members'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Get pending join requests
      tags:
      - circles
  /circles/members/requests/accept:
    put:
      consumes:
      - application/json
      description: Accepts a pending circle join request (admin only)
      parameters:
      - description: Request ID
        in: query
        name: requestId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'res: Join request accepted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request / Circle is full'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: You are not an admin of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error accepting join request'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Accept a join request
      tags:
      - circles
  /circles/members/role:
    put:
      consumes:
      - application/json
      description: Changes the role of a circle member (admin only)
      parameters:
      - description: Role change request
        in: body
        name: role
        required: true
        schema:
          properties:
            memberId:
              type: integer
            role:
              $ref: '#/definitions/circle.Role'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'res: Member role changed successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request / Invalid role / User is not a member
            of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Authentication failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: You are not an admin of this circle'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error changing member role'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWTKeyAuth: []
      - APIKeyAuth: []
      summary: Change member role
      tags:
      - circles
...
securityDefinitions:
  APIKeyAuth:
    description: donetick issued apikey
    in: header
    name: secretkey
    type: apiKey
  JWTKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"